backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
langchain
faiss-cpu
langchain-huggingface